chore: search bar component rename
