+++ components/Navbar.vue
@@ -184,7 +184,7 @@ import BasicImage from '@/components/shared/view/BasicImage.vue'
 import ColorModeButton from '@/components/common/ColorModeButton.vue'
 import Identity from '@/components/identity/IdentityIndex.vue'
 import NavbarProfileDropdown from '@/components/rmrk/Profile/NavbarProfileDropdown.vue'
-import Search from '@/components/search/SearchBar.vue'
+import Search from '@/components/search/Search.vue'
 
 import PrefixMixin from '@/utils/mixins/prefixMixin'
