+++ components/rmrk/Gallery/Gallery.vue
@@ -123,7 +123,7 @@ type GraphResponse = NFTEntitiesWithCount<GraphNFT>
 type SearchedNftsWithMeta = NFTWithCollectionMeta & NFTMetadata
 const components = {
   GalleryCardList: () => import('./GalleryCardList.vue'),
-  Search: () => import('@/components/search/SearchBar.vue'),
+  Search: () => import('@/components/search/Search.vue'),
   Money: () => import('@/components/shared/format/Money.vue'),
   Pagination: () => import('./Pagination.vue'),
   Loader: () => import('@/components/shared/Loader.vue'),
