--- components/search/SearchBarInput.vue
@@ -1,79 +1,0 @@
-<template>
-  <b-autocomplete
-    ref="searchRef"
-    v-model="name"
-    class="gallery-search"
-    placeholder="Search Artwork, Collection..."
-    icon="search"
-    :open-on-focus="showDefaultSuggestions"
-    clearable
-    max-height="500"
-    dropdown-position="bottom"
-    expanded
-    @blur="$emit('blur')"
-    @keydown.native.enter="onEnter">
-    <template #header>
-      <SearchSuggestion
-        :name="name"
-        :show-default-suggestions="showDefaultSuggestions"
-        :query="query"
-        @close="closeDropDown">
-      </SearchSuggestion>
-    </template>
-  </b-autocomplete>
-</template>
-
-<script lang="ts">
-import {
-  Component,
-  Emit,
-  Prop,
-  Ref,
-  VModel,
-  mixins,
-} from 'nuxt-property-decorator'
-import { SearchQuery } from './types'
-import PrefixMixin from '~/utils/mixins/prefixMixin'
-import KeyboardEventsMixin from '~/utils/mixins/keyboardEventsMixin'
-
-@Component({
-  components: {
-    SearchSuggestion: () => import('./SearchSuggestion.vue'),
-  },
-})
-export default class extends mixins(PrefixMixin, KeyboardEventsMixin) {
-  @Prop({ type: Object, required: false }) public query!: SearchQuery
-  @VModel({ type: String }) name!: string
-  @Ref('searchRef') readonly searchRef
-
-  public created() {
-    this.initKeyboardEventHandler({
-      k: this.bindSearchEvents,
-    })
-  }
-
-  @Emit('enter')
-  onEnter() {
-    this.closeDropDown()
-  }
-
-  public focusInput(): void {
-    this.searchRef?.focus()
-  }
-
-  private bindSearchEvents(event) {
-    event.preventDefault()
-    if (event.key === 'k') {
-      this.focusInput()
-    }
-  }
-
-  public closeDropDown() {
-    this.searchRef.isActive = false
-  }
-
-  get showDefaultSuggestions() {
-    return this.urlPrefix === 'rmrk'
-  }
-}
-</script>
